top level acceleration structure, decouple Mesh and BVH

+++ Raytracer/Vertex.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "Math.h"
+#include <xmmintrin.h>
 
 struct Vertex
 {
@@ -7,7 +8,15 @@ struct Vertex
 	Vec3 color_diffuse;
 	Vec3 color_specular;
 	Vec3 normal;
-	double shininess;
+	float shininess;
+	//union {
+	//	Vec3 normal;
+	//	struct {
+	//		float pad0, pad1, pad2, shininess;
+	//	};
+	//	__m128 norm4;
+	//};
+	
 	void Reset() {
 		position = Vec3::Zero();
 		color_diffuse = Vec3::Zero();
